requirements.txt
SQLAlchemy~=2.0.39
python-dotenv~=1.0.1
requests~=2.32.3
pyTelegramBotAPI~=4.26.0
aiogram~=3.18.0
aiohttp~=3.11.14
mysql-connector-python~=9.2.0
pydantic~=2.10.6
redis~=5.2.1